Move "ModInitializer" to the minecraft api package.
